% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bio_loop.r
\name{Bio_loop}
\alias{Bio_loop}
\title{Bio_loop}
\usage{
Bio_loop(
  simyears,
  firstyear,
  dynamic.stocks,
  spawnmonth,
  SSB,
  SSB0,
  NAA,
  ageinfo,
  recrmonths,
  Recr,
  BH_steep,
  recrdev,
  recrfrac,
  NAL,
  ALK,
  Mortalities,
  sexes,
  CALcomm,
  ages,
  Fleets,
  past.trips.by.yearwaveareaboattype,
  areas,
  boat.types,
  subareas,
  trip.types.by.area,
  catch.per.trip.prob.all,
  length.max.legal.all,
  length.min.legal.all,
  length.weight.params,
  lengths,
  modeled.stock.rec.catch.source,
  noncompliance.rate,
  num.keep.legal.all,
  num.keep.legal.group.by.area,
  opt.out.prob.by.area,
  rec.sel.at.length,
  release.mortality.rate,
  stocks.model.pop.this.area,
  stocks.rec.catch.by.area,
  trip.cost.by.area,
  trip.type.utility.coefs,
  utility.catch.weight.coefs,
  utility.coefs,
  utlity.catch.squared.term.names,
  utlity.catch.term.names,
  or.lingcod.bottomfish.catch.per.trip.model,
  wa.lingcod.bottomfish.catch.per.trip.model,
  wa.lingcod.salmon.catch.per.trip.model
)
}
\arguments{
\item{simyears}{Years for simulation}

\item{firstyear}{First year of simulation}

\item{dynamic.stocks}{Names of stocks that change over time}

\item{spawnmonth}{List of spawning months for each stock}

\item{SSB}{Initialized spawning biomass for each stock}

\item{SSB0}{Unfished spawning biomass parameter for each stock}

\item{NAA}{Numbers at age to be simulated for each stock}

\item{ageinfo}{Age information for each stock}

\item{recrmonths}{Months of recruitment for each stock}

\item{Recr}{Recruitment for each stock}

\item{BH_steep}{Recruitment steepness parameter}

\item{recrdev}{Recruitment deviations for Monte Carlo runs}

\item{recrfrac}{Recruitment fraction for sexes}

\item{NAL}{Numbers at length to be simulated for each stock}

\item{ALK}{Age length key}

\item{Mortalities}{Mortalities for each stock}

\item{sexes}{Sexes}

\item{CALcomm}{Commercial catches to be simulated}

\item{ages}{Ages that exist for each stock}

\item{Fleets}{Fleets that exist}

\item{past.trips.by.yearwaveareaboattype}{Historical trip data}

\item{areas}{List of areas to fish}

\item{boat.types}{List of boat types (charter or private)}

\item{subareas}{List of subareas}

\item{trip.types.by.area}{List of trip types by area}

\item{catch.per.trip.prob.all}{Probability of number of fish caught per 
trip}

\item{length.max.legal.all}{Maximum length of retention (by stock)}

\item{length.min.legal.all}{Minimum length of retention (by stock)}

\item{length.weight.params}{Length-weight conversion parameters}

\item{lengths}{Length bins for dynamic stocks}

\item{modeled.stock.rec.catch.source}{Mapping of sub-area stocks to 
dynamic models}

\item{noncompliance.rate}{Rate of non-compliance with retention limits}

\item{num.keep.legal.all}{Max number of fish retained by stock}

\item{num.keep.legal.group.by.area}{Max number of fish retained by 
species group}

\item{opt.out.prob.by.area}{Opt out probability by area}

\item{rec.sel.at.length}{Recreational selectivity at length}

\item{release.mortality.rate}{Release mortality rate by stock}

\item{stocks.model.pop.this.area}{Data loaded for stock population for
this area}

\item{stocks.rec.catch.by.area}{Recreational catches by area}

\item{trip.cost.by.area}{Trip cost by area}

\item{trip.type.utility.coefs}{Previously modeled utility coefficients
for each trip type}

\item{utility.catch.weight.coefs}{Utility coefficients for catch weight}

\item{utility.coefs}{Coefficients for utility model}

\item{utlity.catch.squared.term.names}{Categories for quadratic utility 
terms}

\item{utlity.catch.term.names}{Categories for linear utility terms}

\item{or.lingcod.bottomfish.catch.per.trip.model}{Catch model for Oregon
lingcod that depends on stock status}

\item{wa.lingcod.bottomfish.catch.per.trip.model}{Catch model for WA
lingcod that depends on stock status}

\item{wa.lingcod.salmon.catch.per.trip.model}{Jointcatch model for WA
lingcod and salmon that depends on stock status}
}
\value{
outlist List of stock status and catches
}
\description{
Simulate bioeconomic model across years
}
\examples{

}
