% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_arca_wrapper.r
\name{blast_arca_wrapper}
\alias{blast_arca_wrapper}
\title{blast_arca_wrapper}
\usage{
blast_arca_wrapper(MCMC = 1, firstyear, forecastyrs, rootdir, dynamic.stocks)
}
\arguments{
\item{MCMC}{Number of monte carlo runs per scenario}

\item{firstyear}{First year of simulation}

\item{forecastyrs}{Number of forecast years}

\item{footdir}{Working directory}
}
\value{
dataCompile: a data matrix
}
\description{
Wrapper function for recreational bioeconomic model
}
\examples{

}
