% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateUtility.r
\name{CalculateUtility}
\alias{CalculateUtility}
\title{CalculateUtility}
\usage{
CalculateUtility(
  own.boat,
  opt.out,
  utility.coefs,
  area = NULL,
  sub.area = NULL,
  type = NULL,
  catch.weights = NULL,
  utility.catch.weight.coefs = NULL,
  opt.out.prob = NULL,
  catch.weight.binned = TRUE,
  trip.cost = NULL,
  trip.type.utility.coefs.subarea = NULL,
  utlity.catch.term.names = NULL,
  utlity.catch.squared.term.names = NULL
)
}
\arguments{
\item{own.boat}{Own a boat? 0=no, 1=yes}
}
\value{
utility utility
}
\description{
CalculateUtility
}
\examples{

}
