% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchWeightsEachChoice.r
\name{CatchWeightsEachChoice}
\alias{CatchWeightsEachChoice}
\title{CatchWeightsEachChoice}
\usage{
CatchWeightsEachChoice(catch.lengths.stock.choice, length.weight.params.stock)
}
\arguments{
\item{catch.lengths.stock.choice}{Kept and discarded fish by length}

\item{length.weight.params.stock}{Length-weight parameters}
}
\value{
choice.weights Kept and released weights
}
\description{
Find catch weights
}
\examples{

}
