% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawCatchAll.r
\name{DrawCatchAll}
\alias{DrawCatchAll}
\title{DrawCatchAll}
\usage{
DrawCatchAll(
  catch.per.trip,
  compliant,
  num.keep.legal,
  num.keep.legal.group,
  length.min.legal.all,
  length.max.legal.all,
  rec.sel.at.length,
  lengths,
  len.pop.female,
  len.pop.male,
  modeled.stock.rec.catch.source.subarea
)
}
\arguments{
\item{catch.per.trip}{Number of fish caught}
}
\value{
catch.info catch.info
}
\description{
DrawCatchAll
}
\examples{

}
