% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenLengthAgeKey.r
\name{GenLengthAgeKey}
\alias{GenLengthAgeKey}
\title{GenLengthAgeKey}
\usage{
GenLengthAgeKey(numbersatage, ALKey)
}
\arguments{
\item{numbersatage}{Vector of current numbers at age}
}
\value{
tmp6 tmp6
}
\description{
Organize the numbersatage and age-length-key
}
\examples{

}
