% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_dy_var.r
\name{init_dy_var}
\alias{init_dy_var}
\title{init_dy_var}
\usage{
init_dy_var(dynamic.stocks)
}
\arguments{
\item{dynamic.stocks}{Names of dynamic stocks}
}
\value{
outlist List of stock characteristics
}
\description{
Initialize dynamic variables for stocks
}
\examples{

}
