% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthWeight.r
\name{LengthWeight}
\alias{LengthWeight}
\title{LengthWeight}
\usage{
LengthWeight(length, sex, a.1, b.1, a.2, b.2, a.3, b.3)
}
\arguments{
\item{length}{Fork length in cm}
}
\value{
weight weight
}
\description{
LengthWeight
}
\examples{

}
