% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_res.r
\name{Plot_res}
\alias{Plot_res}
\title{Plot_res}
\usage{
Plot_res(blastres, MCMC, samplei = FALSE)
}
\arguments{
\item{blastres}{Results from running the blast_arca_wrapper}

\item{MCMC}{Number of Monte Carlo iterations}

\item{samplei}{Whether to draw a random iteration or default take the
mean}
}
\value{
plotout Numbers at age plot and rec catches plot
}
\description{
Plot results
}
\examples{

}
