% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_SS.r
\name{Read_SS}
\alias{Read_SS}
\title{Read_SS}
\usage{
Read_SS(filename)
}
\arguments{
\item{filename}{File path where Report.sso is located}
}
\value{
Stockinfo Information about stock used in BLAST model
}
\description{
Read in a Report.sso file from Stock Synthesis for use in BLAST
}
\examples{

}
