% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TripDetails.r
\name{TripDetails}
\alias{TripDetails}
\title{TripDetails}
\usage{
TripDetails(
  curryear,
  currmonth,
  wave,
  area,
  subarea,
  boattype,
  boat.types,
  triptype,
  catch.per.trip.prob.all,
  dynamic.stocks,
  fishing.area.subareas,
  length.max.legal.all,
  length.min.legal.all,
  length.weight.params,
  lengths,
  modeled.stock.rec.catch.source,
  NAL,
  noncompliance.rate,
  num.keep.legal.all,
  num.keep.legal.group.by.area,
  opt.out.prob.by.area,
  rec.sel.at.length,
  release.mortality.rate,
  stocks.model.pop.this.area,
  stocks.rec.catch.by.area,
  trip.cost.by.area,
  trip.type.utility.coefs,
  trip.types.by.area,
  utility.catch.weight.coefs,
  utility.coefs,
  utlity.catch.squared.term.names,
  utlity.catch.term.names,
  or.lingcod.bottomfish.catch.per.trip.model,
  wa.lingcod.bottomfish.catch.per.trip.model,
  wa.lingcod.salmon.catch.per.trip.model
)
}
\arguments{
\item{curryear}{Current year}
}
\value{
dataCompile: a data matrix
}
\description{
Workhorse function used to identify choices and, subsequently, catches 
within a wave-area combination
}
\examples{

}
