% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchWeightsEachChoice.r
\name{CatchWeightsEachChoice}
\alias{CatchWeightsEachChoice}
\title{CatchWeightsEachChoice}
\usage{
CatchWeightsEachChoice(catch.lengths.stock.choice, length.weight.params.stock)
}
\arguments{
\item{catch.lengths.stock.choice}{Kept and discarded fish by length}
}
\value{
choice.weights choice.weights
}
\description{
CatchWeightsEachChoice
}
\examples{

}
