% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawCatchAll.r
\name{DrawCatchAll}
\alias{DrawCatchAll}
\title{DrawCatchAll}
\usage{
DrawCatchAll(
  catch.per.trip,
  compliant,
  num.keep.legal,
  num.keep.legal.group,
  length.min.legal.all,
  length.max.legal.all,
  rec.sel.at.length,
  lengths,
  len.pop.female,
  len.pop.male,
  modeled.stock.rec.catch.source.subarea
)
}
\arguments{
\item{catch.per.trip}{Number of fish caught}

\item{compliant}{Is the trip compliant with regulations (T/F)}

\item{num.keep.legal}{Bag limit for each stock}

\item{num.keep.legal.group}{Bag limit for each species group}

\item{length.min.legal.all}{Minimum legal length for all stocks}

\item{length.max.legal.all}{Maximum legal length for all stocks}

\item{rec.sel.at.length}{Recreational selectivity at length for all 
stocks}

\item{lengths}{Length bins for all stocks}

\item{len.pop.female}{Number of females by length bin for all stocks}

\item{len.pop.male}{Number of males by length bin for all stocks}

\item{modeled.stock.rec.catch.source.subarea}{Subarea for modeled stock}
}
\value{
catch.info Catch information
}
\description{
Draw catch information
}
\examples{

}
