% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenLengthAgeKey.r
\name{GenLengthAgeKey}
\alias{GenLengthAgeKey}
\title{GenLengthAgeKey}
\usage{
GenLengthAgeKey(numbersatage, ALKey)
}
\arguments{
\item{numbersatage}{Vector of current numbers at age}

\item{ALKey}{DF of age-length-conversion w/ C1=LBins, C2=first age}
}
\value{
tmp6 Probability distribution of ages for each length class
}
\description{
Organize the numbersatage and age-length-key
}
\examples{

}
