% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_dy_var.r
\name{Init_dy_var}
\alias{Init_dy_var}
\title{Init_dy_var}
\usage{
Init_dy_var(dynamic.stocks, indat)
}
\arguments{
\item{dynamic.stocks}{Names of dynamic stocks}
}
\value{
outlist List of stock characteristics
}
\description{
Initialize dynamic variables for stocks
}
\examples{

}
